[Refactor] Extract createNetworkedEntity
